移动启动器相关方法
